Move the scheduler to be a part of primitives/nodes
